feat(xpath): Add itertext to XpathItemTreeNodeData
